+++ packages/core/src/repositories/sequelize/index.ts
@@ -1,0 +1,1 @@
+export * from './sequelize-user-modify-crud.repository.base';