+++ services/authentication-service/src/controllers/index.ts
@@ -14,6 +14,7 @@ import {OtpController} from '../modules/auth/otp.controller';
 import {SignupRequestController} from './signup-request.controller';
 import {AzureLoginController} from '../modules/auth/azure-login.controller';
 import {CognitoLoginController} from '../modules/auth/cognito-login.controller';
+import {SamlLoginController} from '../modules/auth/saml-login.controller';
 
 export * from '../modules/auth/login.controller';
 export * from '../modules/auth/logout.controller';
@@ -34,4 +35,5 @@ export const controllers = [
   SignupRequestController,
   AzureLoginController,
   CognitoLoginController,
+  SamlLoginController,
 ];