+++ services/authentication-service/src/modules/auth/saml-login.controller.ts
@@ -1,0 +1,146 @@
+import {inject} from '@loopback/core';
+import {repository} from '@loopback/repository';
+import {
+  get,
+  getModelSchemaRef,
+  HttpErrors,
+  param,
+  post,
+  Request,
+  requestBody,
+  Response,
+  RestBindings,
+} from '@loopback/rest';
+import {
+  CONTENT_TYPE,
+  ILogger,
+  LOGGER,
+  STATUS_CODE,
+  X_TS_TYPE,
+} from '@sourceloop/core';
+import {
+  authenticate,
+  authenticateClient,
+  AuthenticationBindings,
+  AuthErrorKeys,
+  STRATEGY,
+} from 'loopback4-authentication';
+import {authorize} from 'loopback4-authorization';
+import {AuthCodeBindings, AuthCodeGeneratorFn} from '../../providers';
+import {AuthClientRepository} from '../../repositories';
+import {AuthUser} from './models/auth-user.model';
+import {ClientAuthRequest} from './models/client-auth-request.dto';
+import {TokenResponse} from './models/token-response.dto';
+
+const queryGen = (from: 'body' | 'query') => {
+  return (req: Request) => {
+    return {
+      customState: `client_id=${req[from].client_id}`,
+    };
+  };
+};
+export class SamlLoginController {
+  constructor(
+    @repository(AuthClientRepository)
+    public authClientRepository: AuthClientRepository,
+    @inject(LOGGER.LOGGER_INJECT) public logger: ILogger,
+    @inject(AuthCodeBindings.AUTH_CODE_GENERATOR_PROVIDER)
+    private readonly getAuthCode: AuthCodeGeneratorFn,
+  ) {}
+
+  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
+  @authenticate(
+    STRATEGY.SAML,
+    {
+      accessType: 'offline',
+      scope: ['profile', 'email'],
+      authorizationURL: process.env.SAML_URL,
+      callbackURL: process.env.SAML_CALLBACK_URL,
+      clientID: process.env.SAML_CLIENT_ID,
+      clientSecret: process.env.SAML_CLIENT_SECRET,
+      tokenURL: process.env.SAML_TOKEN_URL,
+    },
+    queryGen('body'),
+  )
+  @authorize({permissions: ['*']})
+  @post('/auth/saml', {
+    description: 'POST Call for saml based login',
+    responses: {
+      [STATUS_CODE.OK]: {
+        description: 'Saml Token Response',
+        content: {
+          [CONTENT_TYPE.JSON]: {
+            schema: {[X_TS_TYPE]: TokenResponse},
+          },
+        },
+      },
+    },
+  })
+  async postLoginViaSaml(
+    @requestBody({
+      content: {
+        [CONTENT_TYPE.FORM_URLENCODED]: {
+          schema: getModelSchemaRef(ClientAuthRequest),
+        },
+      },
+    })
+    clientCreds?: ClientAuthRequest, //NOSONAR
+  ): Promise<void> {
+    //do nothing
+  }
+
+  @authenticate(
+    STRATEGY.SAML,
+    {
+      accessType: 'offline',
+      scope: ['profile', 'email'],
+      authorizationURL: process.env.SAML_URL,
+      callbackURL: process.env.SAML_CALLBACK_URL,
+      clientID: process.env.SAML_CLIENT_ID,
+      clientSecret: process.env.SAML_CLIENT_SECRET,
+      tokenURL: process.env.SAML_TOKEN_URL,
+    },
+    queryGen('query'),
+  )
+  @authorize({permissions: ['*']})
+  @get('/auth/saml-auth-redirect', {
+    responses: {
+      [STATUS_CODE.OK]: {
+        description: 'Saml Redirect Token Response',
+        content: {
+          [CONTENT_TYPE.JSON]: {
+            schema: {[X_TS_TYPE]: TokenResponse},
+          },
+        },
+      },
+    },
+  })
+  async samlCallback(
+    @param.query.string('code') code: string, //NOSONAR
+    @param.query.string('state') state: string,
+    @param.query.string('session_state') sessionState: string, //NOSONAR
+    @inject(RestBindings.Http.RESPONSE) response: Response,
+    @inject(AuthenticationBindings.CURRENT_USER)
+    user: AuthUser | undefined,
+  ): Promise<void> {
+    const clientId = new URLSearchParams(state).get('client_id');
+    if (!clientId || !user) {
+      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
+    }
+    const client = await this.authClientRepository.findOne({
+      where: {
+        clientId,
+      },
+    });
+    if (!client?.redirectUrl) {
+      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
+    }
+    try {
+      const token = await this.getAuthCode(client, user);
+      response.redirect(`${client.redirectUrl}?code=${token}`);
+    } catch (error) {
+      this.logger.error(error);
+      throw new HttpErrors.Unauthorized(AuthErrorKeys.InvalidCredentials);
+    }
+  }
+}